Assign each RpcAgent a unique ID, and use ID for sending RPC messages.
Summary: It is not efficient to use a string destination name in every
send. Moreover, when we add RRef later, RpcAgent will frequently check
RRef ownership. It will be slow as well if we have to go though string
comparison every time. This commit assigns each RpcAgent a unique
integer ID. In the Python send API, applications can provide either
destination name or id. If it is a string name, it will be converted to
id by calling the get_worker_id(workerName) API.

ghstack-source-id: 2ce2d17ebd45a1653ada61eee8afc1d43f1b4ac0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/24195